backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
tensorflow
opencv-python-headless
pillow
